Improve mapping; bump version number

+++ F1WM/DatabaseModel/F1WM/News.cs
@@ -14,7 +14,7 @@ namespace F1WM.DatabaseModel
 		public NewsTag MainTag { get; set; }
 		public string Redirect { get; set; }
 		public bool NewsHidden { get; set; }
-		public ushort CommentCount { get; set; }
+		public int CommentCount { get; set; }
 		public bool IsHighlighted { get; set; }
 		public uint Views { get; set; }
 		public uint PosterId { get; set; }
